+++ src/readings/readings.service.ts
@@ -5,6 +5,7 @@ import { Reading } from 'src/shared/entities/reading.entity';
 import { ReadingDto } from 'src/shared/dtos/reading.dto';
 import { Session } from 'src/shared/entities/session.entity';
 import { User } from 'src/shared/entities/user.entity';
+import { from } from 'rxjs';
 
 
 @Injectable()
@@ -30,17 +31,16 @@ export class ReadingsService {
         let reading = null;
         try{
             reading = await this.readingsRepository.findOne(id, { where: { user: user.id } });
+            if (!reading) throw new Error('Bad Request');
         } catch (err) {
             throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
         }
-        if (!reading) throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
         return reading;
     }
 
     async addReading(reading: ReadingDto) {
         let result = null;
         try {
-            //if (reading.id) delete reading.id;
             result = await this.readingsRepository.save(reading);
         } catch (err) {
             throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
@@ -51,7 +51,6 @@ export class ReadingsService {
     async updateReading(id: string, reading: ReadingDto) {
         let result = null;
         try {
-            //if (reading.id) delete reading.id;
             result = await this.readingsRepository.update(id, reading);
         } catch (err) {
             throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
@@ -59,9 +58,12 @@ export class ReadingsService {
         return result;
     }
 
-    async deleteReading(id: string) {
+    async deleteReading(id: string, user: User) {
+        let reading = null;
         let result = null;
         try {
+            reading = await this.readingsRepository.findOne(id, { where: { user: user.id } });
+            if (!reading) throw new Error('Bad Request');
             result = await this.readingsRepository.delete(id);
         } catch (err) {
             throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
@@ -70,13 +72,13 @@ export class ReadingsService {
     }
 
     
-    async findSessionsByReadingId(id: string): Promise<Reading[]> {
+    async findSessionsByReadingId(id: string, user: User): Promise<Reading[]> {
         let sessions = null;
+        let reading = null;
         try {
-            sessions = await this.sessionsRepository
-                .createQueryBuilder('session')
-                .where('session.reading = :id', { id })
-                .getMany();
+            reading = await this.readingsRepository.findOne(id, { where: { user: user.id } });
+            if (!reading) throw new Error('Bad Request');
+            sessions = await this.sessionsRepository.find({ where: { reading: id, user: user.id } });
         } catch(err) {
             throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
         }
